Refactor playerstats into its own file and added first/last games played
Also rounded floating point numbers to at most 2 decimal places.

Now using playerstats.html instead of stats.html for naming consistency.
