+++ templates/playerstats.html
@@ -1,0 +1,28 @@
+{% extends "template.html" %}
+{% block title %} - {{ name }}'s Stats{% end %}
+{% block content %}
+	{% if error is None %}
+		<h1 id="message">{{ name }}</h1>
+		{% if meetupname %}Meetup Name: {{ meetupname }}<br />{% end %}
+		Games played: {{ numgames }}<br />
+		Max score: {{ maxscore }}<br />
+		Min score: {{ minscore }}<br />
+		Average score: {{ avgscore }}<br />
+		Average rank: {{ avgrank }}<br />
+		Average score last 5 games: {{ avgscore5 }}<br />
+		Average rank last 5 games: {{ avgrank5 }}<br />
+		Last played: {{ maxdate }}<br />
+		First played: {{ mindate }}<br />
+		<br />
+		<a class="button" href="/playerhistory/{{ name }}">Game History</a>
+		{% if is_admin %}
+			<form method="post" action="/playerstats/{{ name }}">
+				<input type="text" name="name" value="{{ name }}" />
+				<input type="text" name="meetupname" {% if meetupname %}value="{{ meetupname }}"{% end %} placeholder="Meetup Name (optional)" />
+				<input type="submit" value="Update Names" />
+			</form>
+		{% end %}
+	{% else %}
+		<h1 id="message">{{ error }}</h1>
+	{% end %}
+{% end %}
